+++ CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## 2021-01-02 - Edu
+## Added 
+- **utils/token_handler2.c** Added **eval_token** function to returns if token is PIPE, REDIR, CMD...
 ## Changed
 - Replaced **strcmp** & **modstrcmp** by **libft 's ft_strcmp** (ft_modstrcmp.c to old)
-
+- Rename **ms_parser.c** as **utils/token_handler.c**
+- Moved  **has_pipe_redir_open** function to **utils/token_handler2.c**
 ###	Fixed
 - Fixing **max_pipes_exceeded** exits whithout leaks & also check actual number of pipes (after tokenize)